requirements.txt
# Mint Scan v8 — Python dependencies
# System requirement: sudo apt install python3-tk
# On aarch64 (ARM64): sudo apt install python3-tk libtk8.6

# Core UI — pinned to 5.2.2 for cross-platform stability
# (5.2.2 tested: x86_64, aarch64, Python 3.9-3.12)
customtkinter==5.2.2
darkdetect>=0.8.0

# System monitoring
psutil>=5.9.0
pillow>=9.0.0

# Network
requests>=2.28.0
netifaces>=0.11.0
speedtest-cli>=2.1.3

# Optional — system tray
# pystray>=0.19.0

# Optional — desktop notifications
# plyer>=2.1.0

# Optional — PDF reports
# reportlab>=4.0.0
